Starting to refactor application to use NextJS. This is an incremental update to to the overall effort
