+++ .gitignore
@@ -12,7 +12,7 @@ node_modules/
 *.tsbuildinfo
 .eslintcache
 .changeset/pre.json
-
+.next/
 dist/
 coverage/
 release/
