--- .husky/pre-commit
@@ -1,1 +1,0 @@
-npx --no-install lint-staged
