--- src/components/Favicon/index.tsx
@@ -1,8 +1,0 @@
-import { Helmet } from '@modern-js/runtime/head';
-import imageSrc from './favicon.svg';
-
-export default () => (
-  <Helmet>
-    <link rel="shortcut icon" type="image/x-icon" href={imageSrc} />
-  </Helmet>
-);
