--- src/components/index.ts
@@ -1,13 +1,0 @@
-import Favicon from './Favicon';
-import LocaleLink from './LocaleLink';
-import Logo from './Logo';
-import Banner from './Banner';
-import PageTitle from './PageTitle';
-
-export * from './Auction';
-export * from './Auth';
-export { Favicon, LocaleLink, Logo, Banner };
-export * from './Logo';
-export * from './Banner';
-export { PageTitle };
-export * from './PageTitle';
